Updated all files

+++ api/authors/delete.php
@@ -1,34 +1,44 @@
 <?php
-    // Headers
-    header('Access-Control-Allow-Origin: *');
-    header('Content-Type: application/json');
+    // Additional Headers for DELETE request
     header('Access-Control-Allow-Methods: DELETE');
     header('Access-Control-Allow-Headers: Access-Control-Allow-Headers,Content-Type,Access-Control-Allow-Methods, Authorization,X-Requested-With');
 
-    include_once '../../config/Database.php';
-    include_once '../../models/Author.php';
-
-    // Instantiate DB & Connect
-    $database = new Database();
-    $db = $database->connect();
-
-    // Instantiate blog post object
-    $author = new Author($db);
-
     // Get raw posted data
+    // The POST data will be in json format so need to decode
+    // file_get_contents reads entire file into a string
+    // php://input is a read-only stream that allows you to read raw data from the request body
     $data = json_decode(file_get_contents("php://input"));
+
+    // Make sure $data's id parameter is not empty
+    if (empty($data->id)) {
+        echo json_encode(
+            array('message' => 'Missing Required Parameter')
+        );
+        exit(); 
+    }
+
+    // Make sure the author id exists
+    $authorExists = isValidId($data->id, $author);
+    if (!$authorExists) {
+        // No rows exist with the given author id
+        echo json_encode(
+            array('message' => 'No Quotes Found')
+        );
+        exit();
+    }
 
     // Set ID to update
     $author->id = $data->id;
 
     // Delete post
     if($author->delete()) {
+        // display as json associative array
         echo json_encode(
-            array('message' => 'Post Deleted')
+            array('id' => $author->id)
         );
     } else {
         echo json_encode(
-            array('message' => 'Post Not Deleted')
+            array('message' => 'Author Not Deleted')
         );
     }
